add support for mandatory parameters
Adds two new fields on the @Option annotation to configure whether or not a parameter is required during the instantiation of the builder (to fix a set of immutable parameters) or during the creation of the object to be build (to enforce providing mandatory parameters).

In case of conflicting configuration, e.g. when a default value is specified simultaneously, a warning is logged and the annotation processor gracefully prefers the attributes during instantiation/creation.